Move type PropsPath
