+++ packages/form-builder/src/types.ts
@@ -160,3 +160,22 @@ export interface FormSectionProps extends CommonProps {
 export interface FormSectionToolbarProps extends FormSectionProps {
 	dragHandleProps: DraggableProvidedDragHandleProps;
 }
+
+/**
+ * Creates a dotted path of a nested object property
+ *
+ * type Test = {
+ * 	foo: {
+ * 		bar?: {
+ * 			baz: string;
+ * 		};
+ * 		foo: number;
+ * 	};
+ * 	bar: number;
+ * };
+ *
+ * type T = DeepKeyOf<Test>; // "bar" | "foo.foo" | "foo.bar.baz"
+ */
+export type PropsPath<T extends object> = {
+	[P in keyof T]: T[P] extends object ? `${string & P}` | `${string & P}.${PropsPath<T[P]>}` : `${string & P}`;
+}[T extends any[] ? number & keyof T : keyof T];
